requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
dj-database-url==1.2.0
django-filer==2.2.3
django-js-asset==2.0.0
django-mixins==0.0.10
django-mptt==0.14.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
drf-yasg==1.21.4
easy-thumbnails==2.8.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.1
openapi==1.1.0
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
reportlab==3.6.12
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tzdata==2022.7
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
whitenoise==6.3.0